refactor: AttributeBindingModel in the dataset_generator module
